+++ packaging/docker/samples/local/start.bash
@@ -1,7 +1,7 @@
 #! /bin/bash
 
 #
-# fdb.bash
+# start.bash
 #
 # This source file is part of the FoundationDB open source project
 #
@@ -20,10 +20,20 @@
 # limitations under the License.
 #
 
-source /var/fdb/scripts/create_server_environment.bash
-create_server_environment
-source /var/fdb/.fdbenv
-echo "Starting FDB server on $PUBLIC_IP:$FDB_PORT"
-fdbserver --listen_address 0.0.0.0:$FDB_PORT --public_address $PUBLIC_IP:$FDB_PORT \
-	--datadir /var/fdb/data --logdir /var/fdb/logs \
-	--locality_zoneid=`hostname` --locality_machineid=`hostname` --class $FDB_PROCESS_CLASS
+set -eu
+
+FDB_CLUSTER_FILE="${FDB_CLUSTER_FILE:-docker.cluster}"
+FDB_PORT="${FDB_PORT:-4550}"
+
+FDB_PORT=$FDB_PORT docker-compose up -d fdb
+echo "docker:docker@127.0.0.1:$FDB_PORT" > $FDB_CLUSTER_FILE
+
+# Attempt to connect. Configure the database if necessary.
+if ! fdbcli -C $FDB_CLUSTER_FILE --exec status --timeout 1 ; then
+    if ! fdbcli -C $FDB_CLUSTER_FILE --exec "configure new single memory ; status" --timeout 10 ; then 
+        echo "Unable to configure new FDB cluster."
+        exit 1
+    fi
+fi
+
+echo "Can now connect to docker-based FDB cluster using $FDB_CLUSTER_FILE."